lib/DeviceManager: Add attribute when adding a device and unittest
Adds an attribute (e.g. 'self.lan') with the name of the device type
when adding a new device.

This will raise an exception if you try to add a device of type
that has the same name as an existing attribute (I don't expect that
to ever happen).

Also aliases self.board to self.DUT.

Signed-off-by: Mike Anderson <mike@meldnetworks.com>
Change-Id: If66718f461a4e97d6ca093c09509db8d6916ed77
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/2801
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>